This function is an alternative procedure for finding the optimum (psi, rho) pair. This function first finds all the find all the global optima, and then finds the centroid of this set of globla optima. Then we find the global optimum which is nearest to the centroid. (When the set of global optima is convex, we expect the selected optimum to be at the centroid.)

This function is an alternative procedure for finding the optimum (psi, rho) pair. This function first finds all the find all the global optima, and then finds the centroid of this set of globla optima. Then we find the global optimum which is nearest to the centroid. (When the set of global optima is convex, we expect the selected optimum to be at the centroid.)

find_centroid_of_global_minima(
  d,
  ref_seg_matrix,
  ref_major,
  ref_minor,
  s,
  dist_choice,
  minimise,
  new_bounds,
  distancepng,
  gamma_param,
  siglevel_BAF,
  maxdist_BAF,
  siglevel_LogR,
  maxdist_LogR,
  allow100percent,
  uninformative_BAF_threshold,
  read_depth
)

Arguments

d

A distance matrix

ref_seg_matrix

The corresponding ref seg matrix that belongs to d

ref_major

The corresponding major allele values with d

ref_minor

The corresponding minor allele values with d

s

A segmented BAF/LogR data.frame from get_segment_info

dist_choice

Some distance metrics require adaptation of the data (i.e. log transform)

minimise

Boolean whether we're minimising or maximising

new_bounds

The rho/psi boundaries between we are searching for a solution. This is a named list with values psi_min, psi_max, rho_min, rho_max

distancepng

String where the sunrise distance plot will be saved

gamma_param

The platform gamma

siglevel_BAF

The level at which BAF becomes significant TODO: this option is no longer used

maxdist_BAF

TODO: this option is no longer used

siglevel_LogR

The p-value at which logR becomes significant when establishing whether a segment should be subclonal

maxdist_LogR

The maximum distance allowed as slack when establishing the significance. This allows for the case when a breakpoint is missed, the segment would then not automatically become subclonal

allow100percent

Boolean whether to allow for a 100"%" cellularity solution

uninformative_BAF_threshold

The threshold above which BAF becomes uninformative

read_depth

TODO: this option is no longer used

Value

A list with fields optima_info_without_ref and optima_info