Fit copy number

Function that will fit a clonal copy number profile to segmented data. It first matches the raw LogR with the segmented BAF to create segmented LogR. Then ASCAT is run to obtain a clonal copy number profile. Beyond logRsegmented it produces the rho_and_psi file and the cellularity_ploidy file.

fit.copy.number(
  samplename,
  outputfile.prefix,
  inputfile.baf.segmented,
  inputfile.baf,
  inputfile.logr,
  dist_choice,
  ascat_dist_choice,
  min.ploidy = 1.6,
  max.ploidy = 4.8,
  min.rho = 0.1,
  max.rho = 1,
  min.goodness = 63,
  uninformative_BAF_threshold = 0.51,
  gamma_param = 1,
  use_preset_rho_psi = F,
  preset_rho = NA,
  preset_psi = NA,
  read_depth = 30,
  analysis = "paired",
  nthreads,
  enhanced_grid_search = F
)

Arguments

samplename

Samplename used to name the segmented logr output file

outputfile.prefix

Prefix used for all output file names, except logRsegmented

inputfile.baf.segmented

Filename that points to the BAF segmented data

inputfile.baf

Filename that points to the raw BAF data

inputfile.logr

Filename that points to the raw LogR data

dist_choice

The distance metric that is used internally to rank clonal copy number solutions

ascat_dist_choice

The distance metric used to obtain an initial cellularity and ploidy estimate

min.ploidy

The minimum ploidy to consider (Default 1.6)

max.ploidy

The maximum ploidy to consider (Default 4.8)

min.rho

The minimum cellularity to consider (Default 0.1)

max.rho

The maximum cellularity to consider (Default 1.0)

min.goodness

The minimum goodness of fit for a solution to have to be considered (Default 63)

uninformative_BAF_threshold

The threshold beyond which BAF becomes uninformative (Default 0.51)

gamma_param

Technology parameter, compaction of Log R profiles. Expected decrease in case of deletion in diploid sample, 100 "%" aberrant cells; 1 in ideal case, 0.55 of Illumina 109K arrays (Default 1)

use_preset_rho_psi

Boolean whether to use user specified rho and psi values (Default F)

preset_rho

A user specified rho to fit a copy number profile to (Default NA)

preset_psi

A user specified psi to fit a copy number profile to (Default NA)

read_depth

Legacy parameter that is no longer used (Default 30)

analysis

A String representing the type of analysis to be run, this determines whether the distance figure is produced (Default paired)

Author

dw9, sd11