Correct the LogR estimates for GC content

This function performs GC correction of the LogR data. Sometimes a wave pattern is observed there that correlates with GC content. Internally it uses the ASCAT gc correction function.

gc.correct(
  samplename,
  infile.logr.baf,
  outfile.tumor.LogR,
  outfile.tumor.BAF,
  outfile.normal.LogR,
  outfile.normal.BAF,
  outfile.probeBAF,
  snp6_reference_info_file,
  chr_names,
  birdseed_report_file = "birdseed.report.txt",
  genomebuild = "hg19"
)

Arguments

samplename

Name of the sample to be used to name columns

infile.logr.baf

String that points to the raw combined BAF and LogR file that is the result of cel2baf.logr

outfile.tumor.LogR

The filename of the file where the tumour LogR will be written

outfile.tumor.BAF

The filename of the file where the tumour BAF will be written

outfile.normal.LogR

The filename of the file where the normal LogR will be written

outfile.normal.BAF

The filename of the file where the normal BAF will be written

outfile.probeBAF

The filename of the file where the probe ids and their BAF will be saved

snp6_reference_info_file

String to the SNP6 reference info file that comes with Battenberg SNP6

chr_names

A vector of chromosome names that are to be used

birdseed_report_file

Name of the birdseed output file. This is a temp output file of one of the internally called functions of which the name cannot be defined. Don't change this parameter. (Default birdseed.report.txt)

Author

sd11