Function to correct LogR for waivyness that correlates with GC content

Function to correct LogR for waivyness that correlates with GC content

gc.correct.wgs(
  Tumour_LogR_file,
  outfile,
  correlations_outfile,
  gc_content_file_prefix,
  replic_timing_file_prefix,
  chrom_names,
  recalc_corr_afterwards = F
)

Arguments

Tumour_LogR_file

String pointing to the tumour LogR output

outfile

String pointing to where the GC corrected LogR should be written

correlations_outfile

File where correlations are to be saved

gc_content_file_prefix

String pointing to where GC windows for this reference genome can be found. These files should be split per chromosome and this prefix must contain the full path until chr in its name. The .txt extension is automatically added.

replic_timing_file_prefix

Like the gc_content_file_prefix, containing replication timing info (supply NULL if no replication timing correction is to be applied)

chrom_names

A vector containing chromosome names to be considered

recalc_corr_afterwards

Set to TRUE to recalculate correlations after correction

Author

jdemeul, sd11