Prepare data for impute

Prepare data for impute

generate.impute.input.wgs.germline(
  chrom,
  germline.allele.counts.file,
  normal.allele.counts.file,
  output.file,
  imputeinfofile,
  is.male,
  problemLociFile = NA,
  useLociFile = NA,
  heterozygousFilter = 0.1
)

Arguments

chrom

The chromosome for which impute input should be generated.

germline.allele.counts.file

Output from the allele counter on the matched germline for this chromosome.

normal.allele.counts.file

Output from the allele counter on the matched normal for this chromosome.

output.file

File where the impute input for this chromosome will be written.

imputeinfofile

Info file with impute reference information.

is.male

Boolean denoting whether this sample is male (TRUE), or female (FALSE).

problemLociFile

A file containing genomic locations that must be discarded (optional).

useLociFile

A file containing genomic locations that must be included (optional).

heterozygousFilter

The cutoff where a SNP will be considered as heterozygous (default 0.01).

Author

dw9, sd11, Naser Ansari-Pour (BDI, Oxford)