Obtain BAF and LogR from the allele counts

Obtain BAF and LogR from the allele counts

getBAFsAndLogRs(
  tumourAlleleCountsFile.prefix,
  normalAlleleCountsFile.prefix,
  figuresFile.prefix,
  BAFnormalFile,
  BAFmutantFile,
  logRnormalFile,
  logRmutantFile,
  combinedAlleleCountsFile,
  chr_names,
  g1000file.prefix,
  minCounts = NA,
  samplename = "sample1",
  seed = as.integer(Sys.time())
)

Arguments

tumourAlleleCountsFile.prefix

Prefix of the allele counts files for the tumour.

normalAlleleCountsFile.prefix

Prefix of the allele counts files for the normal.

figuresFile.prefix

Prefix for output figures file names.

BAFnormalFile

File where BAF from the normal will be written.

BAFmutantFile

File where BAF from the tumour will be written.

logRnormalFile

File where LogR from the normal will be written.

logRmutantFile

File where LogR from the tumour will be written.

combinedAlleleCountsFile

File where combined allele counts for tumour and normal will be written.

chr_names

A vector with allowed chromosome names.

g1000file.prefix

Prefix to where 1000 Genomes reference files can be found.

minCounts

Integer, minimum depth required for a SNP to be included (optional, default=NA).

samplename

String, name of the sample (optional, default=sample1).

seed

A seed to be set for when randomising the alleles.

Author

dw9, sd11