Prepare WGS data for haplotype construction

This function performs part of the Battenberg WGS pipeline: Counting alleles, constructing BAF and logR and performing GC content correction.

prepare_wgs(
  chrom_names,
  tumourbam,
  normalbam,
  tumourname,
  normalname,
  g1000allelesprefix,
  g1000prefix,
  gccorrectprefix,
  repliccorrectprefix,
  min_base_qual,
  min_map_qual,
  allelecounter_exe,
  min_normal_depth,
  nthreads,
  skip_allele_counting,
  skip_allele_counting_normal = F
)

Arguments

chrom_names

A vector containing the names of chromosomes to be included

tumourbam

Full path to the tumour BAM file

normalbam

Full path to the normal BAM file

tumourname

Identifier to be used for tumour output files

normalname

Identifier to be used for normal output files

g1000allelesprefix

Prefix path to the 1000 Genomes alleles reference files

g1000prefix

Prefix path to the 1000 Genomes SNP reference files

gccorrectprefix

Prefix path to GC content reference data

repliccorrectprefix

Prefix path to replication timing reference data (supply NULL if no replication timing correction is to be applied)

min_base_qual

Minimum base quality required for a read to be counted

min_map_qual

Minimum mapping quality required for a read to be counted

allelecounter_exe

Path to the allele counter executable (can be found in $PATH)

min_normal_depth

Minimum depth required in the normal for a SNP to be included

nthreads

The number of paralel processes to run

skip_allele_counting

Flag, set to TRUE if allele counting is already complete (files are expected in the working directory on disk)

skip_allele_counting_normal

Flag, set to TRUE from the second sample onwards for multisample case (Default: FALSE)

Author

sd11