Generic reading function using the readr R package, tailored for reading in genomic data

Generic reading function using the readr R package, tailored for reading in genomic data

read_table_generic(
  file,
  header = T,
  row.names = F,
  stringsAsFactor = F,
  sep = "\t",
  chrom_col = 1,
  skip = 0
)

Arguments

file

Filename of the file to read in

header

Whether the file contains a header (Default: TRUE)

row.names

Whether the file contains row names (Default: FALSE)

stringsAsFactor

Legacy parameter that is no longer used (Default: FALSE)

sep

Column separator (Default: \t)

chrom_col

The column number that contains chromosome denominations. This column will automatically be cast as a character. Should be counted including the row.names (Default: 1)

skip

The number of rows to skip before reading (Default: 0)

Value

A data frame with contents of the file