Command to run beagle5

This runs beagle through a system call to the beagle java jar file. It requires pre-formatted reference and plink files for the correct genome build.

run.beagle5(
  beaglejar,
  vcfpath,
  reffile,
  outpath,
  plinkfile,
  nthreads = 1,
  window = 40,
  overlap = 4,
  maxheap.gb = 10,
  javajre = "java"
)

Arguments

beaglejar

character string path to Beagle5 java jar file

vcfpath

character string path to the vcf input file to be phased

reffile

character string path to the Beagle5 reference file

outpath

character string path to Beagle's output vcf.gz file

plinkfile

character string path to the plink file

nthreads

integer number of threads

window

integer max size of genomic window to be phased (cM; default 40; decrease for less memory usage; should be >1.1*overlap)

overlap

integer overlap of windows (cM; default 4)

maxheap.gb

integer maximum heap size for the java process in gigabytes (default 10)

javajre

Path to the Java JRE executable (default java, i.e. in $PATH)

Author

maxime.tarabichi