A modified ASCAT main function to fit Battenberg

This function returns an initial rho and psi estimate for a clonal copy number fit. It uses an internal distance metric to create a distance matrix. Using that matrix it will search for a rho and psi combination that yields the least heavy penalty.

runASCAT(
  lrr,
  baf,
  lrrsegmented,
  bafsegmented,
  chromosomes,
  dist_choice,
  distancepng = NA,
  copynumberprofilespng = NA,
  nonroundedprofilepng = NA,
  cnaStatusFile = "copynumber_solution_status.txt",
  gamma = 0.55,
  allow100percent,
  reliabilityFile = NA,
  min.ploidy = 1.6,
  max.ploidy = 4.8,
  min.rho = 0.1,
  max.rho = 1,
  min.goodness = 63,
  uninformative_BAF_threshold = 0.51,
  chr.names,
  analysis = "paired"
)

Arguments

lrr

(unsegmented) log R, in genomic sequence (all probes), with probe IDs

baf

(unsegmented) B Allele Frequency, in genomic sequence (all probes), with probe IDs

lrrsegmented

log R, segmented, in genomic sequence (all probes), with probe IDs

bafsegmented

B Allele Frequency, segmented, in genomic sequence (only probes heterozygous in germline), with probe IDs

chromosomes

a list containing c vectors, where c is the number of chromosomes and every vector contains all probe numbers per chromosome

dist_choice

The distance metric to be used internally to penalise a copy number solution

distancepng

if NA: distance is plotted, if filename is given, the plot is written to a .png file (Default NA)

copynumberprofilespng

if NA: possible copy number profiles are plotted, if filename is given, the plot is written to a .png file (Default NA)

nonroundedprofilepng

if NA: copy number profile before rounding is plotted (total copy number as well as the copy number of the minor allele), if filename is given, the plot is written to a .png file (Default NA)

cnaStatusFile

File where the copy number profile status is written to. This contains either the message "No suitable copy number solution found" or "X copy number solutions found" (Default copynumber_solution_status.txt)

gamma

technology parameter, compaction of Log R profiles (expected decrease in case of deletion in diploid sample, 100 "%" aberrant cells; 1 in ideal case, 0.55 of Illumina 109K arrays) (Default 0.55)

allow100percent

A boolean whether to allow a 100"%" cellularity solution

reliabilityFile

String to where fit reliabilty information should be written. This file contains backtransformed BAF and LogR values for segments using the fitted copy number profile (Default NA)

min.ploidy

The minimum ploidy to consider (Default 1.6)

max.ploidy

The maximum ploidy to consider (Default 4.8)

min.rho

The minimum cellularity to consider (Default 0.1)

max.rho

The maximum cellularity to consider (Default 1.0)

min.goodness

The minimum goodness of fit for a solution to have to be considered (Default 63)

uninformative_BAF_threshold

The threshold beyond which BAF becomes uninformative (Default 0.51)

chr.names

A vector with chromosome names used for plotting

analysis

A String representing the type of analysis to be run, this determines whether the distance figure is produced (Default paired)

Value

A list with fields psi, rho and ploidy