Plot Battenberg copy number solutions for a segment

squaresplot plots the different Battenberg copy number solutions for a segment

squaresplot(
  tumourname,
  run_dir,
  segment_chr,
  segment_pos,
  platform_gamma = 1,
  pdf = 0,
  binwidth_baf = 0.25,
  xylimits = c(-0.2, 5)
)

Arguments

tumourname

Sample name

run_dir

Running directory

segment_chr

Chromosome containing the segment to be investigated

segment_pos

Chromosomal position within the segment in Mb (e.g. 90M)

platform_gamma

Platform-specific gamma value (0.55 for SNP6, 1 for NGS), default 1

pdf

Output format: 0 for png (default), 1 for pdf

binwidth_baf

BAF isobafline spacing, default 0.25

xylimits

x/y-axis limits, default c(-0.2,5)

Details

The plot is output to the run directory as "tumourname_squares_chr_position.png/pdf"

Author

jd